BUILD: make sure we build on ppc64 and arm64
A few things had crept in that broke arm.

We fix those things and add tests for travis.

We use bb as that is a convenient way to make sure everything builds.

+++ pkg/cpio/sysinfo_linux.go
@@ -14,7 +14,7 @@ func sysInfo(n string, sys *syscall.Stat_t) Info {
 		Mode:     uint64(sys.Mode),
 		UID:      uint64(sys.Uid),
 		GID:      uint64(sys.Gid),
-		NLink:    sys.Nlink,
+		NLink:    uint64(sys.Nlink),
 		MTime:    uint64(sys.Mtim.Sec),
 		FileSize: uint64(sys.Size),
 		Dev:      sys.Dev,
